registration migration: generate the missing sections
there is a typo in the section name resulting in no registration_section_options
being generated.

I also generated an outbound_auth_section_options on all trunk. This was the
behavior in wazo-confgend when translating from sip to pjsip

The username and password from the register are used to generate the
registration outbound auth section