pjsip: global trunk to trunk with registration
rename the template such that the label shows that it's for trunks with
registrations only.

dissociate trunks with no registration from this template and associate them
with the global sip template.

After this migration all trunks should be associated to the global sip template
OR the trunk with registration template (which inherits from global)